requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
monai[all]>=1.3.0
nibabel>=5.1.0
numpy>=1.24.0
scipy>=1.11.0

# Training and evaluation
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
PyYAML>=6.0.1

# Medical imaging
SimpleITK>=2.3.0
scikit-image>=0.21.0
medpy>=0.4.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
